removed deprecated style module
