+++ client/src/components/project-details-plate/utils.js
@@ -1,19 +1,3 @@
-import { styled } from '@mui/material/styles';
-import { 
-  Paper,
-} from '@mui/material';
-
-export const Item = styled(Paper)(({ theme }) => ({
-  backgroundColor: theme.palette.mode === 'dark' ? '#1A2027' : '#fff',
-  padding: theme.spacing(1),
-  textAlign: 'center',
-  color: theme.palette.text.secondary,
-  display: 'flex',
-  alignItems: 'center',
-  minWidth: '345px',
-  height: '56px',
-}));
-
 export const createDetailsItems = (project) => {
   return [
     {
