+++ client/src/components/project-info-table/styles.js
@@ -1,10 +1,1 @@
-import { makeStyles } from '@mui/styles';
 
-export const useStyles = makeStyles(() => {
-  return {
-    wrapper: {
-      border: '1px solid #c1c1c1',
-      borderRadius: '4px',
-    },
-  };
-});
